amend! mf: Handle MESourceSeeked and MEStreamSeeked state change events.
mf: Handle MESourceSeeked and MEStreamSeeked state change events.

If a session is paused then started again, or if a start position is
provided we wait forever for the source to send a MESourceStarted event,
when it sends a MESourceSeeked instead.

This causes a deadlock in Bright Memory: Infinite as the game then waits
for the Start command to complete, while the session is stuck in the
SESSION_STATE_STARTING_SOURCES state.

CW-Bug-Id: #19957
Signed-off-by: Rémi Bernon <rbernon@codeweavers.com>
Signed-off-by: Nikolay Sivov <nsivov@codeweavers.com>
Signed-off-by: Alexandre Julliard <julliard@winehq.org>
(cherry picked from commit 899ae42b206a8c2e462a789249e2c2f143d4a582)